// src/components/Types/users/AuthUsers.ts
interface IUser {
    username: string
    password: string
    avatar: string
    email: string
}

export const users: IUser = {
    username: 'test-user',
    password: 'qwerty',
    avatar: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQQQkPP8eab7uHyMEqDatLnJ7h1fUcOhWWCdQ&usqp=CAU',
    email: "test@email.com"
}